% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edgereduce.R
\name{edgereduce}
\alias{edgereduce}
\title{structure and logic gate inference}
\usage{
edgereduce(data, q1, q2)
}
\arguments{
\item{data}{Data set}

\item{q1}{Threshold for ML}

\item{q2}{Threshold for CMI}
}
\value{
Structure
}
\description{
This function calls structure together with logic gates in gene regulatory networks with un-known structure. For this aim, it takes in steady-state gene expression profiles in (0,1) interval.
}
